requirements.txt
flask==1.0.2
pyjwt==1.7.0
python-dotenv==0.9.1
